pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "attune"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "groq>=0.9.0",
    "click>=8.1",
    "google-api-python-client>=2.120.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[project.scripts]
attune = "attune.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["attune*"]
